+++ .github/workflows/main.yml
@@ -1,24 +1,13 @@
 on: [pull_request] # change to on pullrequest when ready
 
 jobs:
-  hello_world_job:
+  validate_student_ids:
     runs-on: ubuntu-latest
     name: Match PR author with course registered KTH ids
     steps:
     # Checkout latest version of repository
       - name: Checkout
         uses: actions/checkout@v2
-    # test
-      - name: changed files
-        uses: jitterbit/get-changed-files@v1
-        id: changed-files
-        with:
-          # Format of the steps output context.
-          # Can be 'space-delimited', 'csv', or 'json'.
-          # Default: 'space-delimited'
-          format: 'json'
     # Run
       - name: Match PR author with course registered KTH ids
-        uses: ./contributions/course-automation/kalpet-andnil5 # Uses an action in the root directory
-        # with: 
-        #   changed-files: '${{ steps.changed-files.outputs }}'
+        uses: ./contributions/course-automation/kalpet-andnil5
