+++ contributions/course-automation/kalpet-andnil5/src/config/config.js
@@ -1,0 +1,6 @@
+const { join, resolve } = require('path');
+
+module.exports = {
+  KTH_IDS_FILE: join(resolve(__dirname), 'kth-ids.txt'),
+};
+
